+++ src/test/kotlin/com/github/unclepomedev/ronassist/formatting/RonFormatterTestCaseBase.kt
@@ -1,5 +1,6 @@
 package com.github.unclepomedev.ronassist.formatting
 
+import com.intellij.application.options.CodeStyle
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.psi.codeStyle.CodeStyleManager
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
@@ -11,9 +12,24 @@ abstract class RonFormatterTestCaseBase : BasePlatformTestCase() {
     protected fun doTest() {
         val testName = getTestName(false)
         myFixture.configureByFile("$testName/before.ron")
+        val selectionModel = myFixture.editor.selectionModel
         WriteCommandAction.runWriteCommandAction(project) {
-            CodeStyleManager.getInstance(project).reformat(myFixture.file)
+            if (selectionModel.hasSelection()) {
+                CodeStyleManager.getInstance(project).reformatText(
+                    myFixture.file,
+                    selectionModel.selectionStart,
+                    selectionModel.selectionEnd
+                )
+            } else {
+                CodeStyleManager.getInstance(project).reformat(myFixture.file)
+            }
         }
         myFixture.checkResultByFile("$testName/after.ron")
+    }
+
+    protected fun doTestWithTrailingComma(enabled: Boolean) {
+        val settings = CodeStyle.createTestSettings(CodeStyle.getSettings(project))
+        settings.getCustomSettings(RonCodeStyleSettings::class.java).addTrailingComma = enabled
+        CodeStyle.doWithTemporarySettings(project, settings, Runnable { doTest() })
     }
 }